requirements.txt
pandas==2.3.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==3.0.4
joblib==1.3.2
streamlit==1.48.1
matplotlib==3.7.5
seaborn==0.12.2
plotly==6.3.0
requests==2.32.4
beautifulsoup4==4.13.4
rapidfuzz==3.13.0
pandasai==3.0.0b19
pandasai-litellm==0.0.1
litellm==1.79.0
google-generativeai==0.8.5
python-dotenv==1.1.1